Functionality for exit, login, signout
Sign out button is now disabled on app launch, and it will enable when login is successful. When login is successful the login button disables. Clicking exit will now exist the app.

Creating an account on firebase using email, signing out, then signing in with the same email doesn't work. Will need to find a way around that if there is time.